+++ guardrails/telemetry/__init__.py
@@ -1,3 +1,36 @@
-from .telemetry import get_disable_telemetry
+from guardrails.telemetry.common import (
+    default_otel_collector_tracer,
+    default_otlp_tracer,
+    wrap_with_otel_context,
+)
+from guardrails.telemetry.guard_tracing import (
+    trace_guard_execution,
+    trace_async_guard_execution,
+)
+from guardrails.telemetry.open_inference import trace_llm_call, trace_operation
+from guardrails.telemetry.runner_tracing import (
+    trace_step,
+    trace_async_step,
+    trace_stream_step,
+    trace_async_stream_step,
+    trace_call,
+    trace_async_call,
+)
+from guardrails.telemetry.validator_tracing import trace_validator
 
-__all__ = ["get_disable_telemetry"]
+__all__ = [
+    "default_otel_collector_tracer",
+    "default_otlp_tracer",
+    "wrap_with_otel_context",
+    "trace_guard_execution",
+    "trace_async_guard_execution",
+    "trace_llm_call",
+    "trace_operation",
+    "trace_step",
+    "trace_async_step",
+    "trace_stream_step",
+    "trace_async_stream_step",
+    "trace_call",
+    "trace_async_call",
+    "trace_validator",
+]
