+++ guardrails/telemetry/common.py
@@ -1,0 +1,170 @@
+import json
+import sys
+from typing import Any, Callable, Dict, Optional, Union
+from opentelemetry import context
+from opentelemetry.context import Context
+from opentelemetry.trace import Tracer, Span
+
+from guardrails.logger import logger
+from guardrails.stores.context import (
+    get_tracer as get_context_tracer,
+    get_tracer_context,
+)
+
+
+def get_tracer(tracer: Optional[Tracer] = None) -> Optional[Tracer]:
+    # TODO: Do we ever need to consider supporting non-otel tracers?
+    _tracer = tracer if tracer is not None else get_context_tracer()
+    return _tracer
+
+
+def get_current_context() -> Union[Context, None]:
+    otel_current_context = (
+        context.get_current()
+        if context is not None and hasattr(context, "get_current")
+        else None
+    )
+    tracer_context = get_tracer_context()
+    return otel_current_context or tracer_context
+
+
+def get_span(span: Optional[Span] = None) -> Optional[Span]:
+    if span is not None and hasattr(span, "add_event"):
+        return span
+    try:
+        from opentelemetry import trace
+
+        current_context = get_current_context()
+        current_span = trace.get_current_span(current_context)
+        return current_span
+    except Exception as e:
+        logger.error(e)
+        return None
+
+
+def serialize(val: Any) -> Optional[str]:
+    try:
+        if val is None:
+            return None
+        if hasattr(val, "to_dict"):
+            return json.dumps(val.to_dict())
+        elif hasattr(val, "__dict__"):
+            return json.dumps(val.__dict__)
+        elif isinstance(val, dict) or isinstance(val, list):
+            return json.dumps(val)
+        return str(val)
+    except Exception:
+        return None
+
+
+def to_dict(val: Any) -> Dict:
+    try:
+        if val is None:
+            return {}
+        elif isinstance(val, dict):
+            return val
+        elif hasattr(val, "to_dict"):
+            return val.to_dict()
+        elif hasattr(val, "__dict__"):
+            return val.__dict__
+        else:
+            return dict(val)
+    except Exception:
+        return {}
+
+
+def wrap_with_otel_context(
+    outer_scope_otel_context: Context, func: Callable[..., Any]
+) -> Callable[..., Any]:
+    """This function is designed to ensure that a given OpenTelemetry context
+    is applied when executing a specified function. It is particularly useful
+    for preserving the trace context when a guardrails is executed in a
+    different execution flow or when integrating with other frameworks.
+
+    Args:
+        outer_scope_otel_context (Context): The OpenTelemetry context to apply
+            when executing the function.
+        func (Callable[..., Any]): The function to be executed within
+            the given OpenTelemetry context.
+
+    Returns:
+        Callable[..., Any]: A wrapped version of 'func' that, when called,
+            executes with 'outer_scope_otel_context' applied.
+    """
+
+    def wrapped_func(*args: Any, **kwargs: Any) -> Any:
+        # Attach the specified OpenTelemetry context before executing 'func'
+        token = context.attach(outer_scope_otel_context)
+        try:
+            # Execute 'func' within the attached context
+            return func(*args, **kwargs)
+        finally:
+            # Ensure the context is detached after execution
+            #   to maintain correct context management
+            context.detach(token)
+
+    return wrapped_func
+
+
+def default_otel_collector_tracer(resource_name: str = "guardrails"):
+    """This is the standard otel tracer set to talk to a grpc open telemetry
+    collector running on port 4317."""
+
+    from opentelemetry import trace
+    from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
+    from opentelemetry.sdk.resources import SERVICE_NAME, Resource
+    from opentelemetry.sdk.trace import TracerProvider
+    from opentelemetry.sdk.trace.export import BatchSpanProcessor
+
+    resource = Resource(attributes={SERVICE_NAME: resource_name})
+
+    traceProvider = TracerProvider(resource=resource)
+    processor = BatchSpanProcessor(OTLPSpanExporter())
+    traceProvider.add_span_processor(processor)
+    trace.set_tracer_provider(traceProvider)
+
+    return trace.get_tracer("gr")
+
+
+def default_otlp_tracer(resource_name: str = "guardrails"):
+    """This tracer will emit spans directly to an otlp endpoint, configured by
+    the following environment variables:
+
+    OTEL_EXPORTER_OTLP_PROTOCOL
+    OTEL_EXPORTER_OTLP_TRACES_ENDPOINT
+    OTEL_EXPORTER_OTLP_ENDPOINT
+    OTEL_EXPORTER_OTLP_HEADERS
+
+    We recommend using Grafana to collect your metrics. A full example of how to
+    do that is in our (docs)[https://docs.guardrails.com/telemetry]
+    """
+    import os
+
+    from opentelemetry import trace
+    from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
+    from opentelemetry.sdk.resources import SERVICE_NAME, Resource
+    from opentelemetry.sdk.trace import TracerProvider
+    from opentelemetry.sdk.trace.export import (
+        BatchSpanProcessor,
+        ConsoleSpanExporter,
+        SimpleSpanProcessor,
+    )
+
+    envvars_exist = os.environ.get("OTEL_EXPORTER_OTLP_PROTOCOL") and (
+        os.environ.get("OTEL_EXPORTER_OTLP_TRACES_ENDPOINT")
+        or os.environ.get("OTEL_EXPORTER_OTLP_ENDPOINT")
+    )
+
+    resource = Resource(attributes={SERVICE_NAME: resource_name})
+
+    traceProvider = TracerProvider(resource=resource)
+    span_exporter = OTLPSpanExporter()
+    if envvars_exist:
+        processor = BatchSpanProcessor(span_exporter=span_exporter)
+    else:
+        processor = SimpleSpanProcessor(ConsoleSpanExporter(out=sys.stderr))
+
+    traceProvider.add_span_processor(processor)
+    trace.set_tracer_provider(traceProvider)
+
+    return trace.get_tracer("guardrails-ai")
