+++ tests/conftest.py
@@ -9,14 +9,14 @@ os.environ["OPENAI_API_KEY"] = "mocked"
 
 @pytest.fixture(scope="session", autouse=True)
 def mock_tracer():
-    with patch("guardrails.utils.telemetry_utils.get_tracer") as mock_get_tracer:
+    with patch("guardrails.telemetry.common.get_tracer") as mock_get_tracer:
         mock_get_tracer.return_value = None
         yield mock_get_tracer
 
 
 @pytest.fixture(scope="session", autouse=True)
 def mock_span():
-    with patch("guardrails.utils.telemetry_utils.get_span") as mock_get_span:
+    with patch("guardrails.telemetry.common.get_span") as mock_get_span:
         mock_get_span.return_value = None
         yield mock_get_span
